+++ pennylane/templates/subroutines/__init__.py
@@ -54,3 +54,4 @@ from .mod_exp import ModExp
 from .out_poly import OutPoly
 from .gqsp import GQSP
 from .select_pauli_rot import SelectPauliRot
+from .elbow import Elbow
